Create 3Longest_Substring_Without_Repeating
j as fast pointer and i as slow pointer. j keeps moving forward and add non-repeating characters into the set. i moves forward if there is a repeat on the head of the set. 
"abcadef"==> The longest substring is "bcadef". When j==4, the set already contains an "a", remove "a" from the set and recur the loop, i=1 which means the set begins with "b", add "a" which j=4 back to the set.